move GenerationTarget to the SPI package, since it's exposed by the SPI
