+++ hibernate-core/src/main/java/org/hibernate/tool/schema/TargetType.java
@@ -8,7 +8,7 @@ package org.hibernate.tool.schema;
  * Describes the allowable targets (SPI wise) for schema management actions.
  * <p>
  * Under the covers corresponds to provider-specific implementations of
- * {@link org.hibernate.tool.schema.internal.exec.GenerationTarget}
+ * {@link org.hibernate.tool.schema.spi.GenerationTarget}.
  *
  * @author Steve Ebersole
  */
