+++ hibernate-core/src/main/java/org/hibernate/tool/schema/spi/GenerationTarget.java
@@ -2,10 +2,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  * Copyright Red Hat Inc. and Hibernate Authors
  */
-package org.hibernate.tool.schema.internal.exec;
-
-import org.hibernate.tool.schema.spi.SchemaManagementException;
-import org.hibernate.tool.schema.spi.ScriptSourceInput;
+package org.hibernate.tool.schema.spi;
 
 /**
  * Describes a schema generation target
