+++ hibernate-core/src/test/java/org/hibernate/orm/test/schemafilter/RecordingTarget.java
@@ -10,7 +10,7 @@ import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.hibernate.tool.schema.internal.exec.GenerationTarget;
+import org.hibernate.tool.schema.spi.GenerationTarget;
 
 class RecordingTarget implements GenerationTarget {
 	public enum Category {
