+++ hibernate-core/src/test/java/org/hibernate/orm/test/tool/schema/TargetDatabaseImpl.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 import org.hibernate.engine.jdbc.connections.spi.JdbcConnectionAccess;
-import org.hibernate.tool.schema.internal.exec.GenerationTarget;
+import org.hibernate.tool.schema.spi.GenerationTarget;
 import org.hibernate.tool.schema.spi.SchemaManagementException;
 
 /**
